pyproject.toml
[project]
name = "quickscript"
version = "0.2.0"
description = "A concise micro-framework for writing llm-agent scripts."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
]
classifiers = ["Intended Audience :: Developers", "License :: OSI Approved :: Apache Software License", "Natural Language :: English", "Programming Language :: Python :: 3", "Programming Language :: Python :: 3.9", "Programming Language :: Python :: 3.10", "Programming Language :: Python :: 3.11", "Programming Language :: Python :: 3.12", "Programming Language :: Python :: 3.13"]

[dependency-groups]
dev = [
    "pandas>=2.2.3",
    "polars>=1.22.0",
    "pyarrow>=19.0.1",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "pre-commit>=4.1.0",
]
server = [
    "starlette>=0.28.0",
    "uvicorn>=0.27.0",
]
graphql = [
    "starlette>=0.28.0",
    "strawberry-graphql>=0.217.1",
]
cli = [
    "rich>=13.9.4",
]
zmq = [
    "pyzmq>=26.2.1",
]
grpc = [
    "grpcio>=1.70.0",
]
